More steps towards SegWit, now commitment for BTC fee payment is done in advance, test has changed towards BTC, still some changes remaining.

+++ NColorCore/RPC/RPCClient.cs
@@ -110,7 +110,7 @@ namespace NColorCore.RPC
             return txid;
         }
 
-        public async Task<uint256> IssueAssetAsync(BitcoinAddress sourceAddress, BitcoinColoredAddress toAddress, int amount)
+        public async Task<uint256> IssueAssetAsync(BitcoinAddress sourceAddress, BitcoinColoredAddress toAddress, long amount)
         {
             List<object> parameters = new List<object>();
             parameters.Add(sourceAddress.ToString());
